Factor out the metadata makers
